requirements.txt
# XBRL Parsing (Primary)
arelle-release>=2.20.0

# Data Manipulation & Analysis
pandas>=2.0.0
numpy>=1.24.0

# Database & Storage
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pyarrow>=12.0.0

# Web & API
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data Validation & Models
pydantic>=2.0.0
python-dotenv>=1.0.0

# Visualization (Simple)
streamlit>=1.28.0
plotly>=5.17.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
pylint>=3.0.0
mypy>=1.5.0

# Optional: SEC API (if using)
# sec-api>=1.0.0

